requirements.txt
annotated-types==0.7.0
arrow==1.3.0
attrs==25.3.0
aws-cdk-lib==2.208.0
aws-cdk.asset-awscli-v1==2.2.242
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==45.2.0
aws-sam-translator==1.99.0
aws-xray-sdk==2.14.0
aws_lambda_powertools==3.18.0
binaryornot==0.4.4
black==25.1.0
blinker==1.9.0
boto3==1.40.6
botocore==1.40.6
cattrs==24.1.3
certifi==2025.8.3
cffi==1.17.1
cfn-lint==1.36.1
chardet==5.2.0
charset-normalizer==3.4.3
chevron==0.14.0
click==8.1.8
constructs==10.4.2
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==45.0.5
cycler==0.12.1
dateparser==1.2.2
et_xmlfile==2.0.0
fastjsonschema==2.21.1
flake8==7.3.0
fonttools==4.59.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsii==1.112.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.17.0
mypy-boto3-apigateway==1.38.36
mypy-boto3-cloudformation==1.38.31
mypy-boto3-ecr==1.38.37
mypy-boto3-iam==1.38.14
mypy-boto3-kinesis==1.38.8
mypy-boto3-lambda==1.38.40
mypy-boto3-s3==1.38.44
mypy-boto3-schemas==1.38.0
mypy-boto3-secretsmanager==1.38.0
mypy-boto3-signer==1.38.0
mypy-boto3-sqs==1.38.0
mypy-boto3-stepfunctions==1.38.0
mypy-boto3-sts==1.38.38
mypy-boto3-xray==1.38.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
publication==0.0.3
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rpds-py==0.27.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.1
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
text-unidecode==1.3
tomlkit==0.13.3
typeguard==2.13.3
types-awscrt==0.27.4
types-python-dateutil==2.9.0.20250708
types-s3transfer==0.13.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==4.0.2
wheel==0.45.1
wrapt==1.17.2
xlrd==2.0.2
